app/pyproject.toml
[tool.pytest.ini_options]
asyncio_mode="auto"

[tool.poetry]
name = "SHIFT"
version = "0.1.0"
description = ""
authors = ["Tim"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.11.1"
fastapi = "^0.96.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
uvicorn = "^0.22.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^1.10.8"}
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
pytest-cov = "^4.1.0"
isort = "^5.12.0"
sqlalchemy = "^2.0.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
